refactor : RegisterWrapper 삭제
- 중복되는 역할의 RegisterWrapper 삭제
- 해당 역할을 RegisterPage 에서 담당
